Big progress on per-server settings
